[firmware] refs #78 - Register droplet test case in fsm test suite
- TEST_OBJS in Makefile
- Refactor main entrypoint out of test_fsm in to test_main
